refactor: merge spk-solve-validation into spk-solve
spk-solve-validation was the only substantial crate in the workspace with
a single non-dev dependent, and that dependent re-exported it wholesale
(`pub use spk_solve_validation as validation;`). The crate boundary carried
no API meaning -- it was a directory split wearing a crate's clothes.

Its contents (Validators, default_validators, ImpossibleRequestsChecker,
GetMergedRequest) are step-solver internals, so they now live in
crates/spk-solve/src/validation/. Sibling crates graph, solution and
package-iterator each have two or more consumers and are left alone.

The public API is unchanged: spk_solve::validation::* resolves exactly as
before. All 15 source files move byte-identically apart from
spk_solve_validation:: / crate:: -> crate::validation:: path rewrites.

Compile time is unaffected. Three alternating cold builds per configuration
of `cargo build -p spk-solve` (501 units, Linux aarch64, 8 vCPU):

  wall           73.23s -> 71.27s  (-2.7%, sd 6.3/5.6)
  critical path  46.67s -> 46.23s  (-0.9%, sd 4.5/4.3)
  total CPU     321.73s -> 305.60s (-5.0%, sd 23.7/14.4)

Every delta is inside one standard deviation of run-to-run noise. The
critical path is computed from the real cargo unit DAG including rmeta
edges; its shape is unchanged, with the validation node simply absorbed:

  base:   ... graph -> spk-solve-validation (1.58s) -> spk-solve (9.57s)
  merged: ... graph ->                                 spk-solve (16.29s)

Validation was already strictly upstream-serial to spk-solve, so nothing
that previously ran in parallel now serializes.

Tests are conserved exactly: `cargo test -p spk-solve --lib` goes from 370
to 392 passing, which is the 370 it had plus the 22 the validation crate
ran on its own.
