+++ crates/spk-solve/Cargo.toml
@@ -25,6 +25,7 @@ colored = { workspace = true }
 console = { workspace = true }
 crossterm = "0.28.1"
 ctrlc = "3.2"
+dashmap = { workspace = true }
 dyn-clone = { workspace = true }
 enum_dispatch = { workspace = true }
 futures = { workspace = true }
@@ -45,7 +46,6 @@ spk-schema = { workspace = true }
 spk-solve-graph = { workspace = true }
 spk-solve-package-iterator = { workspace = true }
 spk-solve-solution = { workspace = true }
-spk-solve-validation = { workspace = true }
 spk-storage = { workspace = true }
 statsd = { version = "0.15.0", optional = true }
 strum = { workspace = true }
@@ -56,6 +56,7 @@ variantly = { workspace = true }
 
 [dev-dependencies]
 rstest = { workspace = true }
+serde_yaml = { workspace = true }
 spk-solve-macros = { workspace = true }
 strip-ansi-escapes = { workspace = true }
 tap = { workspace = true }
