--- crates/spk-solve/crates/validation/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-authors = { workspace = true }
-edition = { workspace = true }
-name = "spk-solve-validation"
-version = { workspace = true }
-license-file = { workspace = true }
-homepage = { workspace = true }
-repository = { workspace = true }
-readme = { workspace = true }
-description = { workspace = true }
-
-[lints]
-workspace = true
-
-[features]
-
-[dependencies]
-async-trait = { workspace = true }
-colored = { workspace = true }
-dashmap = { workspace = true }
-enum_dispatch = { workspace = true }
-futures = { workspace = true }
-itertools = { workspace = true }
-once_cell = { workspace = true }
-spfs = { workspace = true }
-spk-solve-graph = { workspace = true }
-spk-solve-solution = { workspace = true }
-spk-schema = { workspace = true }
-spk-storage = { workspace = true }
-thiserror = { workspace = true }
-miette = { workspace = true }
-tokio = { workspace = true, features = ["rt"] }
-tracing = "0.1.35"
-
-[dev-dependencies]
-rstest = { workspace = true }
-serde = { workspace = true }
-serde_yaml = { workspace = true }
-spk-solve-macros = { workspace = true }
