+++ crates/spk-solve/src/solvers/resolvo/mod.rs
@@ -41,7 +41,7 @@ use spk_schema::prelude::{HasVersion, Named, Versioned};
 use spk_schema::version_range::VersionFilter;
 use spk_schema::{OptionMap, Package, Spec};
 use spk_solve_solution::{PackageSource, Solution};
-use spk_solve_validation::{Validators, default_validators};
+use crate::validation::{Validators, default_validators};
 use spk_storage::RepositoryHandle;
 
 use crate::solver::Solver as SolverTrait;
