+++ crates/spk-solve/src/validation/error.rs
