+++ crates/spk-solve/src/validation/impossible_checks_test.rs
