+++ crates/spk-solve/src/validation/mod.rs
