+++ crates/spk-solve/src/validation/validation_test.rs
